dom/tv/test/mochitest/test_tv_set_invalid_current_source.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test an error case for SetCurrentSource for TV API</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {
  ok('tv' in navigator, "navigator.tv should exist.");

  navigator.tv.getTuners().then(
    function(aTuners) {
      ok(aTuners.length > 0, "Got at least 1 tuner.");

      aTuners[0].setCurrentSource("InvalidSourceType").then(
        function() {
          ok(false, "Setting an invalid current source should get error.");
          SimpleTest.finish();
        },
        function(aError) {
          is(aError.name, "TypeError", "TypeError should be expected.");
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      SimpleTest.finish();
    }
  );
}

SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
